requirements.txt
streamlit==1.37.1
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0
python-dotenv==1.0.1
requests==2.32.3
